% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_app.R
\name{build_basic_app}
\alias{build_basic_app}
\title{Build basic shiny app}
\usage{
build_basic_app(input_list, datastore_type = "file_based")
}
\arguments{
\item{input_list}{list of fields for the Shiny app}

\item{datastore_type}{type of datastore used to save data. Currently must be file_based}
}
\description{
Build basic shiny CRUD app that stores user input in a csv file
}
\details{
Build a basic shiny app that stores user input in a csv file
}
\examples{
input_list<-list(type=c("Text","Text"),id=c("id1","id2"),label=c("new_blah1","blah2"))
build_basic_app(input_list)
}
